refactor: BoardFactory 클래스 생성
